weather-custom-datamodel/requirements.txt
# This is where required Python libraries are listed
# pandas
# numpy
# scikit-learn
kelvin-app[data]
pytest
